refactor(test): #320 split entity_manager_test.go along service seams
Pure mechanical move (sed range concatenation, no body edited): shared
mocks/fixtures to entity_manager_mocks_test.go; CRUD, query, cross-schema
search, and schema-registry tests to per-concern files. 1196 lines -> five
files, all under the 500-line cap. Test inventory unchanged (diff empty).

Refs #320 #271 #401
